+++ src/connectProvider/web3Provider/web3-provider/providers/ws-provider.js
@@ -34,9 +34,8 @@ class WSProvider {
           '{"jsonrpc":"2.0","method":"net_version","params":[],"id":1}'
         );
       if (
-        /*!Object.is(this.wsProvider, web3.currentProvider) &&*/
-        this.lastMessage + 10 * 60 * 1000 <
-        new Date().getTime() //wait extra 10 minutes
+        !Object.is(this.wsProvider, store.state.web3.currentProvider) &&
+        this.lastMessage + 10 * 60 * 1000 < new Date().getTime() //wait extra 10 minutes
       ) {
         this.wsProvider.disconnect();
         this.oWSProvider.disconnect();